+++ demo/pg_create.go
@@ -13,10 +13,9 @@ import (
 	"strings"
 	"time"
 
-	"github.com/charmbracelet/lipgloss"
+	"github.com/anynines/a9s-cli-v2/makeup"
 	"github.com/fatih/color"
 	"github.com/sethvargo/go-password/password"
-	"golang.org/x/term"
 	"gopkg.in/yaml.v2"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -30,7 +29,7 @@ import (
 // Settings
 // TODO make configurable / cli param
 const configFileName = ".a8s"
-const demoGitRepo = "git@github.com:anynines/a8s-deployment.git"
+const demoGitRepo = "https://github.com/anynines/a8s-deployment.git" // "git@github.com:anynines/a8s-deployment.git"
 const certManagerNamespace = "cert-manager"
 const certManagerManifestUrl = "https://github.com/cert-manager/cert-manager/releases/download/v1.12.0/cert-manager.yaml"
 const defaultDemoSpace = "default"
@@ -39,6 +38,10 @@ const systemName = "a8s Postgres control plane"
 var BackupInfrastructureProvider string // e.g. AWS
 var BackupInfrastructureRegion string   // e.g. us-east-1
 var BackupInfrastructureBucket string   // e.g. a8s-backups
+
+var DeploymentVersion string // e.g. v0.3.0
+var NoPreCheck bool          // e.g. false -> Perform prechecks
+var KubernetesTool string    // e.g. "minikube" or "kind"
 
 // const default_waiting_time_in_s = 10
 
@@ -66,7 +69,7 @@ var DemoConfig Config
 
 func IsCommandAvailable(cmdName string) bool {
 	//	cmd := exec.Command("/bin/sh", "-c", "command -v "+name)
-        //	cmd := exec.Command("command", "-v", cmdName)
+	//	cmd := exec.Command("command", "-v", cmdName)
 	// if err := cmd.Run(); err != nil {
 	path, err := exec.LookPath(cmdName)
 	if err != nil {
@@ -78,12 +81,12 @@ func IsCommandAvailable(cmdName string) bool {
 			msg += " Try running: " + requiredCmds[cmdName][runtime.GOOS]
 		}
 
-		PrintFail(msg)
+		makeup.PrintFail(msg)
 
 		return false
 	}
 
-	PrintCheckmark("Found " + cmdName + " at path " + path + ".")
+	makeup.PrintCheckmark("Found " + cmdName + " at path " + path + ".")
 
 	return true
 }
@@ -92,11 +95,11 @@ func checkIfDockerIsRunning() bool {
 	cmd := exec.Command("docker", "info")
 	err := cmd.Run()
 	if err != nil {
-		PrintFail("Docker is not running.")
-		PrintInfo("Please start the Docker daemon. In case you are using Docker Desktop, start Docker Desktop.")
+		makeup.PrintFail("Docker is not running.")
+		makeup.PrintInfo("Please start the Docker daemon. In case you are using Docker Desktop, start Docker Desktop.")
 		return false
 	}
-	PrintCheckmark("Docker is running.")
+	makeup.PrintCheckmark("Docker is running.")
 	return true
 }
 
@@ -104,12 +107,12 @@ func checkIfKubernetesIsRunning() bool {
 	cmd := exec.Command("kubectl", "api-versions")
 	err := cmd.Run()
 	if err != nil {
-		PrintFail("Kubernetes is not running.")
-		PrintInfo("Please try to restart it or recreate it (delete and re-run the creation).")
-		PrintInfo("Try deleting the Kubernetes cluster with: \"a9s demo delete\". Then recreate it.")
+		makeup.PrintFail("Kubernetes is not running.")
+		makeup.PrintInfo("Please try to restart it or recreate it (delete and re-run the creation).")
+		makeup.PrintInfo("Try deleting the Kubernetes cluster with: \"a9s demo delete\". Then recreate it.")
 		return false
 	}
-	PrintCheckmark("Kubernetes is running.")
+	makeup.PrintCheckmark("Kubernetes is running.")
 	return true
 }
 
@@ -128,17 +131,17 @@ func CheckCommandAvailability() {
 	}
 
 	if !allGood {
-		PrintFailSummary("Sadly, mandatory commands are missing. Aborting...")
+		makeup.PrintFailSummary("Sadly, mandatory commands are missing. Aborting...")
 		os.Exit(1)
 	} else {
-		PrintSuccessSummary("All necessary commands are present.")
+		makeup.PrintSuccessSummary("All necessary commands are present.")
 	}
 }
 
 func CheckPrerequisites() {
 	allGood := true
 
-	PrintH1("Checking Prerequisites...")
+	makeup.PrintH1("Checking Prerequisites...")
 
 	CheckCommandAvailability()
 
@@ -150,68 +153,51 @@ func CheckPrerequisites() {
 		allGood = false
 	}
 
-	if !CheckIfMinkubeClusterExists(DemoClusterName) {
-		CreateMinkubeCluster(DemoClusterName)
+	k8sCreator := GetKubernetesCreator()
+
+	if !k8sCreator.Exists(DemoClusterName) {
+
+		spec := BuildKubernetesClusterSpec()
+
+		k8sCreator.Create(spec, UnattendedMode)
 
 		fmt.Println()
-		PrintH2("Rerunning prerequisite check ...")
+		makeup.PrintH2("Rerunning prerequisite check ...")
 		CheckPrerequisites()
 		allGood = true
+	} else {
+
+		// Only if there's a suitable cluster, the cluster may also be selected.
+		// In any other case, the demo cluster has to be created, first.
+		CheckSelectedCluster()
 	}
 
-	CheckSelectedCluster()
-
-	if !allGood {
-		PrintFailSummary("Sadly, mandatory prerequisited haven't been met. Aborting...")
+	// !NoPreCheck > Perform a pre-check
+	if !NoPreCheck && !allGood {
+		makeup.PrintFailSummary("Sadly, mandatory prerequisited haven't been met. Aborting...")
 		os.Exit(1)
 	}
 }
-func PrintWelcomeScreen() {
-	physicalWidth, _, _ := term.GetSize(int(os.Stdout.Fd()))
-
-	fmt.Println()
-
-	title := "Welcome to the a8s Data Service demos"
-
-	var style = lipgloss.NewStyle().
-		Bold(true).
-		Foreground(lipgloss.Color("#f8f8f8")).
-		Background(lipgloss.Color("#505d78")).
-		PaddingTop(1).
-		PaddingBottom(1).
-		PaddingLeft(0).
-		Width(physicalWidth - 2).
-		Align(lipgloss.Center).
-		//AlignVertical(lipgloss.Top).
-		BorderStyle(lipgloss.RoundedBorder()).
-		BorderForeground(lipgloss.Color("5a6987")).
-		BorderBackground(lipgloss.Color("e4833e"))
-	fmt.Println(style.Render(title))
-
-	PrintH2("This demo will install the a8s Postgres (a8s-pg) demo.")
-
-	WaitForUser()
-}
 
 func EstablishConfigFilePath() {
-	PrintH2("Setting a config file path in order to persist settings...")
+	makeup.PrintH2("Setting a config file path in order to persist settings...")
 
 	homeDir, err := os.UserHomeDir()
 
 	if err != nil {
-		ExitDueToFatalError(err, "Couldn't obtain your home directory. Aborting...")
+		makeup.ExitDueToFatalError(err, "Couldn't obtain your home directory. Aborting...")
 
 	}
 
 	configFilePath = filepath.Join(homeDir, configFileName)
 
-	PrintH2("Settings will be stored at " + configFilePath)
+	makeup.PrintH2("Settings will be stored at " + configFilePath)
 
 }
 
 func EstablishWorkingDir() {
-	PrintH1("Setting up a Working Directory")
-	PrintH2("We will need a working directory for the demo. Let's find one..")
+	makeup.PrintH1("Setting up a Working Directory")
+	makeup.PrintH2("We will need a working directory for the demo. Let's find one..")
 
 	reader := bufio.NewReader(os.Stdin)
 
@@ -219,7 +205,7 @@ func EstablishWorkingDir() {
 		cwd, err := os.Getwd()
 
 		if err != nil {
-			ExitDueToFatalError(err, "Couldn't obtain your current working directory.")
+			makeup.ExitDueToFatalError(err, "Couldn't obtain your current working directory.")
 		}
 
 		fmt.Println("The current working directory is: ", cwd)
@@ -261,13 +247,13 @@ func saveConfig() {
 	yamlData, err := yaml.Marshal(&DemoConfig)
 
 	if err != nil {
-		ExitDueToFatalError(err, "Couldn't save config file. Aborting...")
+		makeup.ExitDueToFatalError(err, "Couldn't save config file. Aborting...")
 	}
 
 	err = os.WriteFile(configFilePath, yamlData, 0600)
 
 	if err != nil {
-		ExitDueToFatalError(err, "Couldn't save config file. Aborting...")
+		makeup.ExitDueToFatalError(err, "Couldn't save config file. Aborting...")
 	}
 }
 
@@ -276,28 +262,22 @@ func LoadConfig() bool {
 
 	if err != nil {
 		if errors.Is(err, os.ErrNotExist) {
-			PrintH2("No config file found.")
+			makeup.PrintH2("No config file found.")
 			return false
 		}
 
-		ExitDueToFatalError(err, "Couldn't open config file.")
+		makeup.ExitDueToFatalError(err, "Couldn't open config file.")
 	}
 
 	err = yaml.Unmarshal(yamlFile, &DemoConfig)
 
 	if err != nil {
-		PrintFail("Coudln't parse config file.")
+		makeup.PrintFail("Coudln't parse config file.")
 	}
 
-	PrintH2("Using the following working directory: " + DemoConfig.WorkingDir)
+	makeup.PrintH2("Using the following working directory: " + DemoConfig.WorkingDir)
 
 	return true
-}
-
-func ExitDueToFatalError(err error, msg string) {
-	PrintFail(msg)
-	fmt.Print(err)
-	os.Exit(1)
 }
 
 func promptPath() string {
@@ -331,22 +311,28 @@ func promptPath() string {
 	}
 }
 
-func CheckoutGitRepository(repositoryURL, localDirectory string) error {
+func CheckoutGitRepository(repositoryURL, localDirectory string, tag string) error {
 	// Check if the local directory already exists
 	if _, err := os.Stat(localDirectory); !os.IsNotExist(err) {
 		return fmt.Errorf("local directory already exists")
 	}
 
-	// Run the git clone command to checkout the repository
-	cmd := exec.Command("git", "clone", repositoryURL, localDirectory)
+	var cmd *exec.Cmd
 
-	PrintCommandBox(cmd.String())
-	WaitForUser()
+	// Run the git clone command to checkout the repository
+	if tag == "latest" {
+		cmd = exec.Command("git", "clone", repositoryURL, localDirectory)
+	} else {
+		cmd = exec.Command("git", "clone", "--branch", tag, repositoryURL, localDirectory)
+	}
+
+	makeup.PrintCommandBox(cmd.String())
+	makeup.WaitForUser(UnattendedMode)
 
 	output, err := cmd.CombinedOutput()
 
 	if err != nil {
-		PrintFail("Failed to checkout the git repository: " + err.Error())
+		makeup.PrintFail("Failed to checkout the git repository: " + err.Error())
 		fmt.Println(string(output))
 		os.Exit(1)
 		return err
@@ -359,34 +345,32 @@ func CheckoutGitRepository(repositoryURL, localDirectory string) error {
 }
 
 func CheckoutDeploymentGitRepository() {
-	PrintH1("Checking out git repository with demo manifests...")
-	Print("Remote Repository is at: " + demoGitRepo)
-	Print("Local working dir: " + DemoConfig.WorkingDir)
-	CheckoutGitRepository(demoGitRepo, DemoConfig.WorkingDir)
+	makeup.PrintH1("Checking out git repository with demo manifests...")
+	makeup.Print("Remote Repository is at: " + demoGitRepo)
+	makeup.Print("Local working dir: " + DemoConfig.WorkingDir)
+	CheckoutGitRepository(demoGitRepo, DemoConfig.WorkingDir, DeploymentVersion)
 }
 
 func CheckSelectedCluster() {
-	Print("Checking whether the " + DemoClusterName + " cluster is selected...")
+	makeup.Print("Checking whether the " + DemoClusterName + " cluster is selected...")
 	cmd := exec.Command("kubectl", "config", "current-context")
 
 	output, err := cmd.CombinedOutput()
 
 	if err != nil {
-		ExitDueToFatalError(err, "Can't retrieve the currently selected cluster using the command: "+cmd.String())
+		makeup.ExitDueToFatalError(err, "Can't retrieve the currently selected cluster using the command: "+cmd.String())
 	}
 
 	current_context := strings.TrimSpace(string(output))
 
-	Print("The currently selected Kubernetes context is: " + current_context)
+	makeup.Print("The currently selected Kubernetes context is: " + current_context)
 
-	var desired_context_name string
-
-	desired_context_name = DemoClusterName
+	desired_context_name := GetKubernetesCreator().GetContext(DemoClusterName)
 
 	if strings.HasPrefix(current_context, desired_context_name) {
-		PrintCheckmark("It seems that the right context is selected: " + desired_context_name)
+		makeup.PrintCheckmark("It seems that the right context is selected: " + desired_context_name)
 	} else {
-		PrintFail("The expected context is " + desired_context_name + " but the current context is: " + current_context + ". Please select the desired context! Try executing: ")
+		makeup.PrintFail("The expected context is " + desired_context_name + " but the current context is: " + current_context + ". Please select the desired context! Try executing: ")
 		fmt.Println("kubectl config use-context " + desired_context_name)
 		os.Exit(1)
 	}
@@ -395,13 +379,13 @@ func CheckSelectedCluster() {
 func GetKubernetesConfigPath() string {
 	var kubeconfig string
 	if kubeconfig = os.Getenv("KUBECONFIG"); kubeconfig != "" {
-		Print("Kubernetes configuration is set by the $KUBECONFIG env variable.")
+		makeup.Print("Kubernetes configuration is set by the $KUBECONFIG env variable.")
 	} else if home := homedir.HomeDir(); home != "" {
-		Print("Kubernetes configuration is set by $HOME/.kube/config.")
+		makeup.Print("Kubernetes configuration is set by $HOME/.kube/config.")
 		flag.CommandLine = flag.NewFlagSet("kubeconfig", flag.ExitOnError)
 		kubeconfig = *flag.String("kubeconfig", filepath.Join(home, ".kube", "config"), "(optional) absolute path to the kubeconfig file")
 	} else {
-		Print("Kubernetes configuration is set by config flag.")
+		makeup.Print("Kubernetes configuration is set by config flag.")
 		kubeconfig = *flag.String("kubeconfig", "", "absolute path to the kubeconfig file")
 	}
 
@@ -417,7 +401,7 @@ func CountPodsInDemoNamespace() int {
 
 func GetKubernetesClientSet() *kubernetes.Clientset {
 	kubeconfig := GetKubernetesConfigPath()
-	Print("Kubernetes config located at: " + kubeconfig)
+	makeup.Print("Kubernetes config located at: " + kubeconfig)
 
 	// use the current context in kubeconfig
 	config, err := clientcmd.BuildConfigFromFlags("", kubeconfig)
@@ -437,7 +421,7 @@ func GetKubernetesClientSet() *kubernetes.Clientset {
 // https://github.com/kubernetes/client-go/blob/master/examples/in-cluster-client-configuration/main.go
 func countPodsInNamespace(namespace string) int {
 
-	PrintH2("Checking whether there are pods in the cluster...")
+	makeup.PrintH2("Checking whether there are pods in the cluster...")
 
 	clientset := GetKubernetesClientSet()
 
@@ -456,11 +440,11 @@ func KubectlApplyF(yamlFilepath string) {
 
 	output, err := cmd.CombinedOutput()
 
-	PrintCommandBox(cmd.String())
-	WaitForUser()
+	makeup.PrintCommandBox(cmd.String())
+	makeup.WaitForUser(UnattendedMode)
 
 	if err != nil {
-		ExitDueToFatalError(err, "Can't kubectl apply with command: "+cmd.String())
+		makeup.ExitDueToFatalError(err, "Can't kubectl apply with command: "+cmd.String())
 	}
 
 	fmt.Println(string(output))
@@ -470,39 +454,39 @@ func KubectlApplyKustomize(kustomizeFilepath string) {
 
 	cmd := exec.Command("kubectl", "apply", "--kustomize", kustomizeFilepath)
 
-	PrintCommandBox(cmd.String())
-	WaitForUser()
+	makeup.PrintCommandBox(cmd.String())
+	makeup.WaitForUser(UnattendedMode)
 
 	output, err := cmd.CombinedOutput()
 
 	fmt.Println(string(output))
 
 	if err != nil {
-		ExitDueToFatalError(err, "Can't kubectl kustomize with using the command: "+cmd.String())
+		makeup.ExitDueToFatalError(err, "Can't kubectl kustomize with using the command: "+cmd.String())
 	}
 
 }
 
 func ApplyA8sManifests() {
-	PrintH1("Applying the a8s Data Service manifests...")
+	makeup.PrintH1("Applying the a8s Data Service manifests...")
 	kustomizePath := filepath.Join(DemoConfig.WorkingDir, "deploy", "a8s", "manifests")
 	KubectlApplyKustomize(kustomizePath)
-	PrintCheckmark("Done applying a8s manifests.")
+	makeup.PrintCheckmark("Done applying a8s manifests.")
 }
 
 func WaitForCertManagerToBecomeReady() {
-	PrintH1("Waiting for the cert-manager API to become ready.")
+	makeup.PrintH1("Waiting for the cert-manager API to become ready.")
 	crashLoopBackoffCount := 10
 
 	for i := 1; i <= crashLoopBackoffCount; i++ {
 		cmd := exec.Command("cmctl", "check", "api")
 		output, err := cmd.CombinedOutput()
 
-		Print(cmd.String())
+		makeup.Print(cmd.String())
 
 		//TODO Crash loop detection / timeout
 		if err != nil {
-			PrintWait("Continuing to wait for the cert-manager API...")
+			makeup.PrintWait("Continuing to wait for the cert-manager API...")
 		}
 
 		strOutput := string(output)
@@ -510,24 +494,24 @@ func WaitForCertManagerToBecomeReady() {
 		fmt.Println(strOutput)
 
 		if strings.TrimSpace(strOutput) == "The cert-manager API is ready" {
-			PrintCheckmark("The cert-manager is ready")
+			makeup.PrintCheckmark("The cert-manager is ready")
 			return
 		} else {
-			PrintWait("Continuing to wait for the cert-manager API...")
+			makeup.PrintWait("Continuing to wait for the cert-manager API...")
 		}
 
 		time.Sleep(30 * time.Second)
 	}
 
-	PrintFailSummary("The cert-manager did not become ready within reasonable time.")
+	makeup.PrintFailSummary("The cert-manager did not become ready within reasonable time.")
 }
 
 func ApplyCertManagerManifests() {
-	PrintH1("Installing the cert-manager")
+	makeup.PrintH1("Installing the cert-manager")
 	count := countPodsInNamespace(certManagerNamespace)
 
 	if count > 0 {
-		Print(fmt.Sprintf("Found %d pods in the %s namespace", count, certManagerNamespace))
+		makeup.Print(fmt.Sprintf("Found %d pods in the %s namespace", count, certManagerNamespace))
 	}
 
 	KubectlApplyF(certManagerManifestUrl)
@@ -560,13 +544,13 @@ Generates an encryption password file for backups if it doesnt exist.
 Does nothing if the file already exists.
 */
 func EstablishEncryptionPasswordFile() {
-	PrintH2("In order to encrypt backups we need an encryption password.")
-	Print("Checking if encryption password file for backups already exists...")
+	makeup.PrintH2("In order to encrypt backups we need an encryption password.")
+	makeup.Print("Checking if encryption password file for backups already exists...")
 
 	filePath := BackupConfigEncryptionPasswordFilePath()
 
 	if _, err := os.Stat(filePath); !os.IsNotExist(err) {
-		Print("There's already an encryption password file. Skipping password generation...")
+		makeup.Print("There's already an encryption password file. Skipping password generation...")
 		return
 	}
 
@@ -575,7 +559,7 @@ func EstablishEncryptionPasswordFile() {
 	backupPassword, err := password.Generate(64, 10, 10, false, false)
 
 	if err != nil {
-		ExitDueToFatalError(err, "Couldn't generate encryption password for backup config.")
+		makeup.ExitDueToFatalError(err, "Couldn't generate encryption password for backup config.")
 	}
 
 	saveStringToFile(filePath, backupPassword)
@@ -591,7 +575,7 @@ func saveStringToFile(filePath, content string) {
 	f, err := os.Create(filePath)
 
 	if err != nil {
-		ExitDueToFatalError(err, "Couldn't create file to store  encryption password for backup config to filepath: "+filePath)
+		makeup.ExitDueToFatalError(err, "Couldn't create file to store  encryption password for backup config to filepath: "+filePath)
 	}
 
 	defer f.Close()
@@ -599,7 +583,7 @@ func saveStringToFile(filePath, content string) {
 	f.WriteString(content)
 
 	if err != nil {
-		ExitDueToFatalError(err, "Couldn't write password to file to store encryption password for backup config to filepath: "+filePath)
+		makeup.ExitDueToFatalError(err, "Couldn't write password to file to store encryption password for backup config to filepath: "+filePath)
 	}
 
 	f.Sync()
@@ -613,7 +597,7 @@ Skips if the file is already present
 func ReadStringFromFileOrConsole(filePath, contentType string, showContent bool) {
 
 	if _, err := os.Stat(filePath); !os.IsNotExist(err) {
-		Print("There's already an " + contentType + " file...")
+		makeup.Print("There's already an " + contentType + " file...")
 		return
 	}
 
@@ -624,11 +608,11 @@ func ReadStringFromFileOrConsole(filePath, contentType string, showContent bool)
 	accessKeyId, err := reader.ReadString('\n')
 
 	if err != nil {
-		ExitDueToFatalError(err, "Can't read  "+contentType+"  from STDIN.")
+		makeup.ExitDueToFatalError(err, "Can't read  "+contentType+"  from STDIN.")
 	}
 
 	if showContent {
-		Print(contentType + " : " + accessKeyId)
+		makeup.Print(contentType + " : " + accessKeyId)
 	}
 
 	// Write file
@@ -641,7 +625,7 @@ If not it prompts to read the access key id from STDIN.
 Skips if the access key id file is already present
 */
 func EstablishAccessKeyId() {
-	PrintH2("In order to store backups on an object store such as S3, we need an ACCESS KEY ID.")
+	makeup.PrintH2("In order to store backups on an object store such as S3, we need an ACCESS KEY ID.")
 
 	filePath := BackupConfigAccessKeyIdFilePath()
 
@@ -649,7 +633,7 @@ func EstablishAccessKeyId() {
 }
 
 func establishSecretAccessKey() {
-	PrintH2("In order to store backups on an object store such as S3, we need a SECRET KEY.")
+	makeup.PrintH2("In order to store backups on an object store such as S3, we need a SECRET KEY.")
 
 	filePath := BackupConfigSecretAccessKeyFilePath()
 
@@ -661,14 +645,14 @@ func backupStoreConfigFilePath() string {
 }
 
 func establishBackupStoreConfigYaml() {
-	PrintH2("Checking the backup-store-config.yaml file...")
+	makeup.PrintH2("Checking the backup-store-config.yaml file...")
 
 	filePath := backupStoreConfigFilePath()
 
 	if CheckIfFileExists(filePath) {
-		PrintCheckmark(fmt.Sprintf("There's already a backup-store-config.yaml file at %s. Trusting that the file is ok.", filePath))
+		makeup.PrintCheckmark(fmt.Sprintf("There's already a backup-store-config.yaml file at %s. Trusting that the file is ok.", filePath))
 	} else {
-		Print("Writing a backup-store-config.yaml with defaults to " + filePath)
+		makeup.Print("Writing a backup-store-config.yaml with defaults to " + filePath)
 
 		// TODO Make backup store configurable
 		blobStoreConfig := BlobStore{
@@ -684,13 +668,13 @@ func establishBackupStoreConfigYaml() {
 		yamlData, err := yaml.Marshal(&blobStoreConfig)
 
 		if err != nil {
-			ExitDueToFatalError(err, "Couldn't generate backup-store-config.yaml file. Aborting...")
+			makeup.ExitDueToFatalError(err, "Couldn't generate backup-store-config.yaml file. Aborting...")
 		}
 
 		err = os.WriteFile(filePath, yamlData, 0644)
 
 		if err != nil {
-			ExitDueToFatalError(err, "Couldn't save backup-store-config.yaml file. Aborting...")
+			makeup.ExitDueToFatalError(err, "Couldn't save backup-store-config.yaml file. Aborting...")
 		}
 	}
 }
@@ -720,30 +704,30 @@ func checkIfPodHasStatusRunningInNamespace(podNameStartsWith string, namespace s
 
 	for _, pod := range pods.Items {
 		if strings.HasPrefix(pod.Name, podNameStartsWith) {
-			Print("Found pod with prefix " + podNameStartsWith)
+			makeup.Print("Found pod with prefix " + podNameStartsWith)
 
 			// if debug {
 			// 	//pod.Status.Phase
-			// 	Print("Pod has status: " + pod.Status.String())
+			// 	makeup.Print("Pod has status: " + pod.Status.String())
 			// }
 
 			switch phase := pod.Status.Phase; phase {
 			case v1.PodRunning:
-				PrintCheckmark("The Pod " + pod.Name + "h is running as expected.")
+				makeup.PrintCheckmark("The Pod " + pod.Name + "h is running as expected.")
 				return true
 			case v1.PodFailed:
-				PrintFail("The Pod " + pod.Name + "h has failed but should be running.")
-				PrintFail("The " + systemName + " has not been installed successfully.")
+				makeup.PrintFail("The Pod " + pod.Name + "h has failed but should be running.")
+				makeup.PrintFail("The " + systemName + " has not been installed successfully.")
 				os.Exit(1)
 
 			case v1.PodPending:
-				Print("The Pod " + pod.Name + "h in pending but should be running.")
+				makeup.Print("The Pod " + pod.Name + "h in pending but should be running.")
 				return false
 			case v1.PodSucceeded:
-				Print("The Pod " + pod.Name + "h has succeeded but should be running.")
+				makeup.Print("The Pod " + pod.Name + "h has succeeded but should be running.")
 				return false
 			case v1.PodUnknown:
-				Print("The Pod " + pod.Name + "h has an unknown status but should be running.")
+				makeup.Print("The Pod " + pod.Name + "h has an unknown status but should be running.")
 				return false
 			default:
 				return false
@@ -754,7 +738,7 @@ func checkIfPodHasStatusRunningInNamespace(podNameStartsWith string, namespace s
 }
 
 func WaitForSystemToBecomeReady() {
-	PrintH1("Waiting for the " + systemName + " to become ready...")
+	makeup.PrintH1("Waiting for the " + systemName + " to become ready...")
 
 	allGood := true
 
@@ -774,37 +758,37 @@ out:
 		// We start optimistically that all pods are running
 		allGood = true
 		for _, expectedPodPrefix := range expectedPodPrefixes {
-			Print("Checking the " + expectedPodPrefix.name + "...")
+			makeup.Print("Checking the " + expectedPodPrefix.name + "...")
 			if checkIfPodHasStatusRunningInNamespace(expectedPodPrefix.name, systemNamespace) {
-				PrintCheckmark("The " + expectedPodPrefix.name + " appears to be running.")
+				makeup.PrintCheckmark("The " + expectedPodPrefix.name + " appears to be running.")
 				expectedPodPrefix.running = true
 			} else {
 				// Sadly, at least one pod isn't running so we need another loop iteration
-				PrintFail("The " + expectedPodPrefix.name + " is not ready (yet).")
+				makeup.PrintFail("The " + expectedPodPrefix.name + " is not ready (yet).")
 				allGood = false
 			}
 
 			if allGood {
-				PrintSuccessSummary("The " + systemName + " appears to be ready. All expected pods are running.")
+				makeup.PrintSuccessSummary("The " + systemName + " appears to be ready. All expected pods are running.")
 				break out
 			} else {
-				PrintWait("The " + systemNamespace + " is not ready (yet), let's try again in 5s ...")
+				makeup.PrintWait("The " + systemNamespace + " is not ready (yet), let's try again in 5s ...")
 				time.Sleep(5 * time.Second)
 			}
 		}
 	}
-	WaitForUser()
+	makeup.WaitForUser(UnattendedMode)
 }
 
 func PrintDemoSummary() {
-	PrintH1("Summary")
-	Print("You've successfully accomplished the followings steps:")
-	PrintCheckmark("Created a Kubernetes Cluster with Kind named: " + DemoClusterName + ".")
-	PrintCheckmark("Installed cert-manager on the Kubernetes cluster.")
-	PrintCheckmark("Created a configuration for the backup object store.")
-	PrintCheckmark("Installing the a8s Postgres control plane.\n")
+	makeup.PrintH1("Summary")
+	makeup.Print("You've successfully accomplished the followings steps:")
+	makeup.PrintCheckmark("Created a Kubernetes Cluster using " + KubernetesTool + " named: " + DemoClusterName + ".")
+	makeup.PrintCheckmark("Installed cert-manager on the Kubernetes cluster.")
+	makeup.PrintCheckmark("Created a configuration for the backup object store.")
+	makeup.PrintCheckmark("Installing the a8s Postgres control plane.\n")
 
 	//TODO Check whether Pods- from the a8s-system are ready
-	//PrintCheckmark("Installed the a8s Postgres control plane.\n")
-	PrintSuccessSummary("You are now ready to create a8s Postgres service instances.")
+	//makeup.PrintCheckmark("Installed the a8s Postgres control plane.\n")
+	makeup.PrintSuccessSummary("You are now ready to create a8s Postgres service instances.")
 }